fixers/deprecation-hunter/requirements.txt
langchain
langchain-openai
openai
gitpython
pygithub
rich
typer
pytest
ast-grep-py
